% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{change_NA}
\alias{change_NA}
\title{Change NAs in a data frame to something else, or zero by default.}
\usage{
change_NA(df, to = 0)
}
\arguments{
\item{df}{The data frame}

\item{to}{The new value for NAs, defaults to zero}
}
\value{
The updated data frame
}
\description{
Change NAs in a data frame to something else, or zero by default.
}
\examples{
df \%>\% change_NA(to=0)
df \%>\% change_NA(to="")
}

