% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{file_purge}
\alias{file_purge}
\title{Removes all but the most recent file with a given pattern, i.e. a directory purge a la VMS.
Beware of using this indiscriminately.}
\usage{
file_purge(path, pattern = "*")
}
\arguments{
\item{path}{The path to the folder containing the files}

\item{pattern}{A regular expression for the files to remove}
}
\value{
The return value from the \code{file.remove} call.
}
\description{
Removes all but the most recent file with a given pattern, i.e. a directory purge a la VMS.
Beware of using this indiscriminately.
}

