% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{find_latest}
\alias{find_latest}
\title{}
\usage{
find_latest(srcDir = ".", pattern)
}
\arguments{
\item{srcDir}{The search directory.}

\item{pattern}{A regular expression matching the file you want to find.}
}
\value{
The full path of the matching file, or NULL.
}
\description{

}
\section{File functions}{


Find the latest version of a file with a given pattern in a given folder.
This is typically used for files that start with a ymd date or a monotonically
increasing string, e.g. \code{yyyy-mm-dd filename.ext}.
}

