% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{lastQ}
\alias{lastQ}
\title{A string representing the previous quarter in the form \code{yyyy-Qq}.}
\format{\preformatted{ chr "2015-Q1"
}}
\usage{
lastQ
}
\description{
A string representing the previous quarter in the form \code{yyyy-Qq}.
}
\keyword{datasets}

