% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{norm}
\alias{norm}
\title{Normalise and round to a given number of digits.}
\usage{
norm(v, digits = 2)
}
\arguments{
\item{v}{A vector}

\item{digits}{The number of digits for rounding, defaults to 2.}
}
\value{
The normalised vector.
}
\description{
Normalise and round to a given number of digits.
}

