% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{read_latest}
\alias{read_latest}
\title{Read the latest CSV or XLSX file with a given pattern in a given folder.}
\usage{
read_latest(srcDir = ".", pattern, skip = 0, sheet = 1,
  R.identifiers = FALSE)
}
\arguments{
\item{srcDir}{The search directory}

\item{pattern}{A regular expression matching the file you want to find}

\item{skip}{The number of lines to skip in the input file, defaults to 0}

\item{sheet}{The sheet number if reading a apreadsheet, defaults to 1}

\item{R.identifiers}{Convert the column names to valid R identifiers, defaults to \code{FALSE}.}
}
\value{
A data frame with the colnames either as exact strings matching the header
values, or as valid R identifies for compatibility with \code{read.csv()}.
}
\description{
Read the latest CSV or XLSX file with a given pattern in a given folder.
}
\examples{
read_latest(src_dir(thisQ), ".*abcdef.*\\\\.txt"))
read_latest(src_dir(thisQ), ".*abcdef.*\\\\.xlsx", sheet=2, R.identifiers=TRUE))
}
\seealso{
\code{\link{find_latest}}
}

