% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{str_qtr}
\alias{str_qtr}
\title{util.R
Useful utility functions for my data wrangling}
\usage{
str_qtr(d)
}
\arguments{
\item{d}{A date value}
}
\value{
A string of the form \code{yyyy-Qq} where \code{q} is one of 1, 2, 3 or 4.
}
\description{
util.R
Useful utility functions for my data wrangling
}
\section{Set up data sources}{



Create a quarter folder using the provided date
}
\examples{
str_qtr(now())
str_qtr(as.Date("2003-03-02"))
}

