% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{thisQ}
\alias{thisQ}
\title{A string representing the current quarter in the form \code{yyyy-Qq}.}
\format{\preformatted{ chr "2015-Q2"
}}
\usage{
thisQ
}
\description{
A string representing the current quarter in the form \code{yyyy-Qq}.
}
\keyword{datasets}

