% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{view_as_html}
\alias{view_as_html}
\title{Show a table as ugly HTML in the Viewer pane}
\usage{
view_as_html(tbl)
}
\arguments{
\item{tbl}{A data frame for printing}
}
\value{
Nothing
}
\description{
Show a table as ugly HTML in the Viewer pane
}

