% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{xnonzero}
\alias{xnonzero}
\title{Select the first value unless zero, in which case return the second!}
\usage{
xnonzero(a, b)
}
\description{
Select the first value unless zero, in which case return the second!
}

