% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ajworkr.R
\docType{package}
\name{ajworkr}
\alias{ajworkr}
\alias{ajworkr-package}
\title{ajworkr: my assorted utility functions}
\description{
This provides a number of utility functions used for reading data from files
and wrangling data.
}

