% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{src_dir}
\alias{src_dir}
\title{A source folder for data usiny my convention for where it's stored.}
\usage{
src_dir(qtr, folder = "Reports", base_dir = "..")
}
\description{
A source folder for data usiny my convention for where it's stored.
}

