% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{str_qtr}
\alias{str_qtr}
\title{util.R
Useful utility functions for my data wrangling}
\usage{
str_qtr(d = today())
}
\arguments{
\item{d}{A date value. Defaults to the current date.}
}
\value{
A string of the form \code{yyyy-Qq} where \code{q} is one of 1, 2, 3 or 4.
}
\description{
Create a string representing the quarter, using the provided date
}
\examples{
str_qtr()
str_qtr("2003-03-02")
}

